+++ services/StateMap.ts
@@ -1,5 +1,5 @@
 import { strict as assert } from 'assert';
-import { StageLinqValue } from '../network';
+import { StageLinqValue } from '../types';
 import { ReadContext } from '../utils/ReadContext';
 import { WriteContext } from '../utils/WriteContext';
 import { Service } from './Service';
