+++ utils/tcp.ts
@@ -1,6 +1,6 @@
 import { Socket as TCPSocket } from 'net';
 import { PromiseSocket } from 'promise-socket';
-import { CONNECT_TIMEOUT } from '../network';
+import { CONNECT_TIMEOUT } from '../types';
 import { Logger } from '../LogEmitter';
 
 export type Connection = PromiseSocket<TCPSocket>;
